Fix load_module raising error and not unloading
It now checks if the handlers/senders are set before trying to add them,
so that nothing from a module is loaded until all functions it is trying
to set are verified to not already exist.

Also fixed a bug where it would raise an error while setting a function
that was already set, because it didn't catch the error.