Add callback `IRCe.DISCONNECT`, hooks, ...
...change callback behaviour.

This is for the host application to call when the server disconnects or
the socket is closed. The application can use this to do cleanup.

Modules can now specify hooks, which are called when a command is
received (this includes `IRCe.DISCONNECT`). These don't affect handlers
or callbacks.

These two points address issue #7.

Callbacks are now called if they exist regardless of whether the handler
returns anything or not. The previous system of not calling a callback
when the handler didn't return anything was a bit overcomplicated and
seemed unecessary. I'm not sure if this breaks anything, we'll see.

Also gave the RAW handler the special value `IRCe.RAW`. Using the string
"RAW" is still included for now for backwards compatibility, but will
be removed at some point.