Revert to not calling callback if handler returns nothing
Y'know what, I think this is the best. Rationale is in README.md,
reproduced here:

This behaviour allows for handlers which sometimes pass processing
on to another handler; eg. when a CTCP ACTION is received, the CTCP
handler passes processing over to ACTION. This way, the ACTION
callback is called instead of the CTCP callback. If the callback
were always called (as they were in versions 2.1.0 to 5.0.0), the
CTCP handler would be called regardless. NOTICE and PRIVMSG also
work this way, passing over to CTCP when they detect a CTCP message.